backend/requirements.txt
crewai==0.203.1
crewai-tools==0.76.0
langchain==0.3.27
langchain-openai==0.3.35
langchain-community==0.3.31
langchain_ibm==0.3.2
langchain_chroma==0.2.6
ibm-watsonx-ai==1.1.26
ibm-watson-machine-learning==1.0.360
unstructured==0.18.15
databricks-sdk==0.68.0
requests
pandas
pydantic
openai
fastapi
uvicorn
python-multipart
python-dotenv
pyyaml
